Removed SystemDriver::fopen method, outdated by the new virtual file system. Respective port maintainers will have to remove cssys/*/fopen.cpp from their project files... I did it already for ports that use the Makefile system. Removed FOpen and FClose (huh?) methods from ISystem interface, for same reason. Removed SaveFile() and LoadFile() methods from csCamera class. Debug methods moved to debug program (i.e. walktest). MazeD now is compilable. However, some important functions are disabled (saving and loading). Many small fixes. Added "const" here and there through the entire CS source tree.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@1622 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgfxldr/walimage.cpp
@@ -59,7 +59,7 @@ ImageWALFile::~ImageWALFile ()
 ImageWALFile::ImageWALFile (UByte* ptr, long filesize) : ImageFile ()
 {
   status=IFE_BadFormat;
-  unsigned int chkfilesize;
+  long chkfilesize;
   unsigned char *datPtr;
   unsigned int palindex;
   unsigned int i;
@@ -69,8 +69,8 @@ ImageWALFile::ImageWALFile (UByte* ptr, long filesize) : ImageFile ()
   
   // There's no id-tag in .WAL files, so the only way I know to check
   // if it's a wal, is to use this method. Hope it works
-  chkfilesize=head->offsets[3]+((head->height/8)*(head->width/8));
-  if(chkfilesize!=filesize) return;
+  chkfilesize = head->offsets[3]+((head->height/8)*(head->width/8));
+  if (chkfilesize != filesize) return;
 
   //There are 4 mipmaps in a wal-file, but we just use the first and
   //Discard the rest
